+++ pkg/proxy/ranges/byterange/multipart.go
@@ -156,7 +156,8 @@ func (mbrs MultipartByteRanges) Compress() {
 }
 
 // ParseMultipartRangeResponseBody returns a MultipartByteRanges from the provided body
-func ParseMultipartRangeResponseBody(body io.Reader, contentTypeHeader string) (MultipartByteRanges, string, Ranges, int64, error) {
+func ParseMultipartRangeResponseBody(body io.Reader,
+	contentTypeHeader string) (MultipartByteRanges, string, Ranges, int64, error) {
 	parts := make(MultipartByteRanges)
 	ranges := make(Ranges, 0)
 	fullContentLength := int64(-1)
@@ -201,12 +202,13 @@ func ParseMultipartRangeResponseBody(body io.Reader, contentTypeHeader string) (
 			}
 		}
 	}
-	return parts, ct, ranges, int64(fullContentLength), nil
+	return parts, ct, ranges, fullContentLength, nil
 }
 
 // ExtractResponseRange returns http headers and body representing the subject MultipartByteRanges map,
 // cropped to the provided ranges
-func (mbrs MultipartByteRanges) ExtractResponseRange(ranges Ranges, fullContentLength int64, contentType string, body []byte) (http.Header, []byte) {
+func (mbrs MultipartByteRanges) ExtractResponseRange(ranges Ranges, fullContentLength int64,
+	contentType string, body []byte) (http.Header, []byte) {
 
 	if ranges == nil || len(ranges) == 0 {
 		return nil, body
